+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - The `craft.blitz.getTemplate` tag now renders the template using the same site as the request originated from.
 
 ### Fixed
+- Fixed an error that could occur if a site defined in the Cloudflare Zone ID settings was deleted ([#192](https://github.com/putyourlightson/craft-blitz/issues/192)).
 - Fixed the table alignment in Blitz Cache Utility when the Yii Cache Storage was used.
 
 ## 3.5.2 - 2020-03-25